+++ extensions/coffee/src/caffeinate.ts
@@ -3,5 +3,11 @@ import { startCaffeinate, getSchedule, changeScheduleState } from "./utils";
 export default async () => {
   const schedule = await getSchedule();
   if (schedule != undefined) await changeScheduleState("decaffeinate", schedule);
-  await startCaffeinate({ menubar: true, status: true }, "Your Mac is now caffeinated");
+  
+  const caffeinationInfo = {
+    type: "manual" as const,
+    startTime: Date.now(),
+  };
+  
+  await startCaffeinate({ menubar: true, status: true }, "☕ Your computer is now caffeinated!", undefined, caffeinationInfo);
 };