+++ extensions/coffee/src/decaffeinate.ts
@@ -4,6 +4,6 @@ import { showToast, Toast } from "@raycast/api";
 export default async () => {
   const schedule = await getSchedule();
   if (schedule != undefined && schedule.IsRunning == true)
-    await showToast(Toast.Style.Failure, "Caffeination schedule running, pause to decaffeinate");
-  else await stopCaffeinate({ menubar: true, status: true }, "Your Mac is now decaffeinated");
+    await showToast(Toast.Style.Failure, "⏸️ Caffeination schedule is running - pause it to decaffeinate");
+  else await stopCaffeinate({ menubar: true, status: true }, "💤 Your computer is now decaffeinated");
 };